refactor(dashboard): replace monolithic FiltersContext with focused FilterProvider architecture
Replace the 1084-line monolithic FiltersContext with a new FilterProvider that splits
functionality into focused hooks (useFilterActionsOnly, useFiltersFull, useFilterStateOnly,
useStatusFilter). This architectural improvement reduces prop drilling, improves performance
through targeted re-renders, and provides better separation of concerns.

- Remove monolithic FiltersContext.tsx (1084 lines)
- Update components to use specific filter hooks instead of generic useFilters
- Simplify component interfaces by removing unused props
- Maintain all existing functionality while improving code organization
- Add proper JSDoc documentation to simplified components
